+++ tsup.config.ts
@@ -31,7 +31,8 @@ import { defineConfig } from 'tsup'
 export default defineConfig([
   {
     entry: {
-      index: 'src/index.ts'
+      index: 'src/index.ts',
+      'browser/index': 'src/browser/index.ts'
     },
     format: ['esm', 'cjs'],
     dts: true,